+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Lint issues (ignored returns)
+- Cluster app namespace if using constructors random org
 
 ### Chore